taskBack/requirements.txt
Django~=3.1.6
djangorestframework~=3.12.2
djangorestframework-jwt==1.11.0